requirements.txt
2captcha-python==1.5.0
aenum==3.1.15
aiohappyeyeballs==2.4.3
appdirs==1.4.4
asgiref==3.8.1
atomicwrites==1.4.0
attrs==24.2.0
bkcharts==0.2
Deprecated==1.2.14
dj-database-url==0.5.0
Django==5.1.2
django-admin-interface==0.28.9
django-allauth==65.0.1
django-colorfield==0.11.0
django-cors-headers==4.5.0
django-environ==0.8.1
django-extensions==3.2.3
django-heroku==0.3.1
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fontawesomefree==6.6.0
future==0.18.3
gunicorn==20.1.0
inflection==0.5.1
jsonpointer==2.1
mysqlclient==2.1.1
oauthlib==3.2.2
outcome==1.3.0.post0
Pillow==10.0.0
protobuf==3.20.3
psycopg2==2.9.5
pyasn1-modules==0.2.8
python-decouple==3.8
python-slugify==8.0.4
python3-openid==3.2.0
requests
scikit-surprise==1.1.4
selenium
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
TwoCaptcha==0.0.1
typing_extensions==4.12.2
webencodings==0.5.1
whitenoise==6.7.0
wsproto==1.2.0
setuptools>=70.0.0
scikit-learn==1.4.2
beautifulsoup4
pandas
python-dateutil
line-bot-sdk==3.14.0
mariadb
openai
python-dotenv
aiohttp>=3.10.9